aca/src/requirements.txt
mcp[cli]
newsapi-python
python-dotenv>=1.0.0